+++ packages/core/src/types/PlatePlugin/OverrideProps.ts
@@ -1,0 +1,6 @@
+import { GetNodeProps } from '../PlatePluginOptions/GetNodeProps';
+import { SPEditor } from '../SPEditor';
+
+export type OverrideProps<T extends SPEditor = SPEditor> = (
+  editor: T
+) => GetNodeProps;